remove vmtransaction
