--- src/main/java/org/tron/common/runtime/vm/program/VMTransaction.java
@@ -1,546 +1,0 @@
-/*
- * Copyright (c) [2016] [ <ether.camp> ]
- * This file is part of the ethereumJ library.
- *
- * The ethereumJ library is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * The ethereumJ library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
- */
-package org.tron.common.runtime.vm.program;
-
-import static org.apache.commons.lang3.ArrayUtils.isEmpty;
-
-import com.google.common.primitives.Longs;
-import java.math.BigInteger;
-import java.security.SignatureException;
-import java.util.Arrays;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.spongycastle.util.BigIntegers;
-import org.spongycastle.util.encoders.Hex;
-import org.tron.common.crypto.ECKey;
-import org.tron.common.crypto.ECKey.ECDSASignature;
-import org.tron.common.crypto.ECKey.MissingPrivateKeyException;
-import org.tron.common.crypto.Hash;
-import org.tron.common.runtime.vm.DataWord;
-import org.tron.common.utils.ByteUtil;
-import org.tron.common.utils.Sha256Hash;
-import org.tron.core.capsule.TransactionCapsule;
-import org.tron.protos.Protocol;
-
-public class VMTransaction {
-
-  private static final Logger logger = LoggerFactory.getLogger("VMTransaction");
-  private static final BigInteger DEFAULT_GAS_PRICE = new BigInteger("10000000000000");
-  private static final BigInteger DEFAULT_BALANCE_GAS = new BigInteger("21000");
-
-  public static final int HASH_LENGTH = 32;
-  public static final int ADDRESS_LENGTH = 20;
-
-  /* SHA3 hash of the RLP encoded transaction */
-  private byte[] hash;
-
-  /* the amount of ether to transfer (calculated as wei) */
-  private long value;
-
-  /* the address of the destination account
-   * In creation transaction the receive address is - 0 */
-  private byte[] receiveAddress;
-
-  /* An unlimited size byte array specifying
-   * input [data] of the message call or
-   * Initialization code for a new contract */
-  private byte[] data;
-
-  /**
-   * Since EIP-155, we could encode chainId in V
-   */
-  private static final int CHAIN_ID_INC = 35;
-  private static final int LOWER_REAL_V = 27;
-  private Integer chainId = null;
-
-  /* the elliptic curve signature
-   * (including public key recovery bits) */
-  private ECDSASignature signature;
-
-  protected byte[] sendAddress;
-
-  /* Tx in encoded form */
-  protected byte[] protoEncoded;
-  private byte[] protoRaw;
-  /* Indicates if this transaction has been parsed
-   * from the RLP-encoded data */
-  protected boolean parsed = false;
-
-  /////////////////////////////
-  Protocol.Transaction trx;
-
-  public VMTransaction(byte[] rawData) {
-    this.protoEncoded = rawData;
-    parsed = false;
-  }
-
-  public VMTransaction(Protocol.Transaction tx) {
-    this.trx = tx;
-    TransactionCapsule trxCap = new TransactionCapsule(tx);
-    this.protoEncoded = trxCap.getData();
-    parsed = false;
-  }
-
-
-  public VMTransaction(byte[] receiveAddress, long value, byte[] data,
-      Integer chainId) {
-    this.receiveAddress = receiveAddress;
-    if (value == 0L) {
-      this.value = 0L;
-    } else {
-      this.value = value;
-    }
-    this.data = data;
-    this.chainId = chainId;
-
-    if (receiveAddress == null) {
-      this.receiveAddress = ByteUtil.EMPTY_BYTE_ARRAY;
-    }
-
-    parsed = true;
-  }
-
-  public VMTransaction(byte[] receiveAddress, long value, byte[] data) {
-    this(receiveAddress, value, data, null);
-  }
-
-  public VMTransaction(byte[] receiveAddress, long value, byte[] data,
-      byte[] r, byte[] s, byte v, Integer chainId) {
-    this(receiveAddress, value, data, chainId);
-    this.signature = ECDSASignature.fromComponents(r, s, v);
-  }
-
-  public VMTransaction(byte[] receiveAddress, long value, byte[] data,
-      byte[] r, byte[] s, byte v) {
-    this(receiveAddress, value, data, r, s, v, null);
-  }
-
-
-  private Integer extractChainIdFromV(BigInteger bv) {
-    if (bv.bitLength() > 31) {
-      return Integer.MAX_VALUE; // chainId is limited to 31 bits, longer are not valid for now
-    }
-    long v = bv.longValue();
-    if (v == LOWER_REAL_V || v == (LOWER_REAL_V + 1)) {
-      return null;
-    }
-    return (int) ((v - CHAIN_ID_INC) / 2);
-  }
-
-  private byte getRealV(BigInteger bv) {
-    if (bv.bitLength() > 31) {
-      return 0; // chainId is limited to 31 bits, longer are not valid for now
-    }
-    long v = bv.longValue();
-    if (v == LOWER_REAL_V || v == (LOWER_REAL_V + 1)) {
-      return (byte) v;
-    }
-    byte realV = LOWER_REAL_V;
-    int inc = 0;
-    if ((int) v % 2 == 0) {
-      inc = 1;
-    }
-    return (byte) (realV + inc);
-  }
-
-  public long transactionCost(Protocol.Block block) {
-
-    protoParse();
-
-    //return config.getConfigForBlock(block.getNumber()).
-    //        getTransactionCost(this);
-    return 0;
-  }
-
-  public synchronized void verify() {
-    protoParse();
-    validate();
-  }
-
-  public synchronized void protoParse() {
-    if (parsed) {
-      return;
-    }
-    try {
-      //RLPList decodedTxList = RLP.decode2(rlpEncoded);
-      //RLPList transaction = (RLPList) decodedTxList.get(0);
-
-      // Basic verification
-            /*if (transaction.size() > 9 ) throw new RuntimeException("Too many RLP elements");
-            for (RLPElement rlpElement : transaction) {
-                if (!(rlpElement instanceof RLPItem))
-                    throw new RuntimeException("Transaction RLP elements shouldn't be lists");
-            }
-
-            this.gasPrice = transaction.get(1).getRLPData();
-            this.gasLimit = transaction.get(2).getRLPData();
-            this.receiveAddress = transaction.get(3).getRLPData();
-            this.value = transaction.get(4).getRLPData();
-            this.data = transaction.get(5).getRLPData();
-            // only parse signature in case tx is signed
-            if (transaction.get(6).getRLPData() != null) {
-                byte[] vData =  transaction.get(6).getRLPData();
-                BigInteger v = ByteUtil.bytesToBigInteger(vData);
-                this.chainId = extractChainIdFromV(v);
-                byte[] r = transaction.get(7).getRLPData();
-                byte[] s = transaction.get(8).getRLPData();
-                this.signature = ECDSASignature.fromComponents(r, s, getRealV(v));
-            } else {
-                logger.debug("RLP encoded tx is not signed!");
-            }
-            this.parsed = true;
-            this.hash = getHash();
-            */
-
-      ///
-
-      //use protoEncoded directly and it should never be null
-      this.hash = Hash.sha3(protoEncoded);
-      this.parsed = true;
-      TransactionCapsule transaction = new TransactionCapsule(protoEncoded);
-    } catch (Exception e) {
-      throw new RuntimeException("Error on parsing proto", e);
-    }
-  }
-
-  private void validate() {
-    if (receiveAddress != null && receiveAddress.length != 0
-        && receiveAddress.length != ADDRESS_LENGTH) {
-      throw new RuntimeException("Receive address is not valid");
-    }
-
-    if (getSignature() != null) {
-      if (BigIntegers.asUnsignedByteArray(signature.r).length > HASH_LENGTH) {
-        throw new RuntimeException("Signature R is not valid");
-      }
-      if (BigIntegers.asUnsignedByteArray(signature.s).length > HASH_LENGTH) {
-        throw new RuntimeException("Signature S is not valid");
-      }
-      if (getSender() != null && getSender().length != ADDRESS_LENGTH) {
-        throw new RuntimeException("Sender is not valid");
-      }
-    }
-  }
-
-  public boolean isParsed() {
-    return parsed;
-  }
-
-  public byte[] getHash() {
-    if (!isEmpty(hash)) {
-      return Arrays.copyOf(hash, hash.length);
-    }
-
-    protoParse();
-    byte[] plainMsg = this.getEncoded();
-    return Hash.sha3(plainMsg);
-  }
-
-  public byte[] getRawHash() {
-    protoParse();
-    byte[] plainMsg = this.getEncodedRaw();
-    return Hash.sha3(plainMsg);
-  }
-
-  public long getValue() {
-    protoParse();
-    return value;
-  }
-
-  protected void setValue(long value) {
-    this.value = value;
-    parsed = true;
-  }
-
-  public byte[] getReceiveAddress() {
-    protoParse();
-    return receiveAddress;
-  }
-
-  protected void setReceiveAddress(byte[] receiveAddress) {
-    this.receiveAddress = receiveAddress;
-    parsed = true;
-  }
-
-
-  public long nonZeroDataBytes() {
-    if (data == null) {
-      return 0;
-    }
-    int counter = 0;
-    for (final byte aData : data) {
-      if (aData != 0) {
-        ++counter;
-      }
-    }
-    return counter;
-  }
-
-  public long zeroDataBytes() {
-    if (data == null) {
-      return 0;
-    }
-    int counter = 0;
-    for (final byte aData : data) {
-      if (aData == 0) {
-        ++counter;
-      }
-    }
-    return counter;
-  }
-
-
-  public byte[] getData() {
-    protoParse();
-    return data;
-  }
-
-  protected void setData(byte[] data) {
-    this.data = data;
-    parsed = true;
-  }
-
-  public ECDSASignature getSignature() {
-    protoParse();
-    return signature;
-  }
-
-  public byte[] getContractAddress() {
-    if (!isContractCreation()) {
-      return null;
-    }
-    // return HashUtil.calcNewAddr(this.getSender(), this.getNonce());
-    DataWord addr = new DataWord(this.getSender());
-    byte[] privKey = Sha256Hash.hash(addr.getData());
-    ECKey ecKey = ECKey.fromPrivate(privKey);
-    byte[] newAddress = ecKey.getAddress();
-    return newAddress;
-  }
-
-  public boolean isContractCreation() {
-    protoParse();
-    return this.receiveAddress == null || Arrays
-        .equals(this.receiveAddress, ByteUtil.EMPTY_BYTE_ARRAY);
-  }
-
-  /*
-   * Crypto
-   */
-
-  public ECKey getKey() {
-    byte[] hash = getRawHash();
-    return ECKey.recoverFromSignature(signature.v, signature, hash);
-  }
-
-  public synchronized byte[] getSender() {
-    try {
-      if (sendAddress == null && getSignature() != null) {
-        sendAddress = ECKey.signatureToAddress(getRawHash(), getSignature());
-      }
-      return sendAddress;
-    } catch (SignatureException e) {
-      logger.error(e.getMessage(), e);
-    }
-    return null;
-  }
-
-  public Integer getChainId() {
-    protoParse();
-    return chainId == null ? null : (int) chainId;
-  }
-
-  /**
-   * @deprecated should prefer #sign(ECKey) over this method
-   */
-  public void sign(byte[] privKeyBytes) throws MissingPrivateKeyException {
-    sign(ECKey.fromPrivate(privKeyBytes));
-  }
-
-  public void sign(ECKey key) throws MissingPrivateKeyException {
-    this.signature = key.sign(this.getRawHash());
-    this.protoEncoded = null;
-  }
-
-  @Override
-  public String toString() {
-    return toString(Integer.MAX_VALUE);
-  }
-
-  public String toString(int maxDataSize) {
-    protoParse();
-    String dataS;
-    if (data == null) {
-      dataS = "";
-    } else if (data.length < maxDataSize) {
-      dataS = ByteUtil.toHexString(data);
-    } else {
-      dataS = ByteUtil.toHexString(Arrays.copyOfRange(data, 0, maxDataSize)) +
-          "... (" + data.length + " bytes)";
-    }
-    return "TransactionData [" + "hash=" + ByteUtil.toHexString(hash) +
-        ", receiveAddress=" + ByteUtil.toHexString(receiveAddress) +
-        ", sendAddress=" + ByteUtil.toHexString(getSender()) +
-        ", value=" + value +
-        ", data=" + dataS +
-        ", signatureV=" + (signature == null ? "" : signature.v) +
-        ", signatureR=" + (signature == null ? ""
-        : ByteUtil.toHexString(BigIntegers.asUnsignedByteArray(signature.r))) +
-        ", signatureS=" + (signature == null ? ""
-        : ByteUtil.toHexString(BigIntegers.asUnsignedByteArray(signature.s))) +
-        "]";
-  }
-
-  /**
-   * For signatures you have to keep also RLP of the transaction without any signature data
-   */
-  public byte[] getEncodedRaw() {
-    protoParse();
-    if (protoRaw != null) {
-      return Arrays.copyOf(protoRaw, protoRaw.length);
-    }
-
-    return this.protoEncoded.clone();
-        /*
-        // parse null as 0 for nonce
-        byte[] gasPrice = RLP.encodeElement(this.gasPrice);
-        byte[] gasLimit = RLP.encodeElement(this.gasLimit);
-        byte[] receiveAddress = RLP.encodeElement(this.receiveAddress);
-        byte[] value = longToBytes(this.value);
-        byte[] data = RLP.encodeElement(this.data);
-
-        // Since EIP-155 use chainId for v
-        if (chainId == null) {
-            protoRaw = RLP.encodeList(nonce, gasPrice, gasLimit, receiveAddress,
-                    value, data);
-        } else {
-            byte[] v, r, s;
-            v = RLP.encodeInt(chainId);
-            r = RLP.encodeElement(EMPTY_BYTE_ARRAY);
-            s = RLP.encodeElement(EMPTY_BYTE_ARRAY);
-            rlpRaw = RLP.encodeList(nonce, gasPrice, gasLimit, receiveAddress,
-                    value, data, v, r, s);
-        }
-        return rlpRaw;
-        */
-  }
-
-  public byte[] getEncoded() {
-
-    if (protoEncoded != null) {
-      if (null == this.hash) {
-        this.hash = Hash.sha3(protoEncoded);
-      }
-      return protoEncoded;
-    }
-        /*
-        byte[] gasPrice = RLP.encodeElement(this.gasPrice);
-        byte[] gasLimit = RLP.encodeElement(this.gasLimit);
-        byte[] receiveAddress = RLP.encodeElement(this.receiveAddress);
-        byte[] value = RLP.encodeElement(this.value);
-        byte[] data = RLP.encodeElement(this.data);
-
-        byte[] v, r, s;
-
-        if (signature != null) {
-            int encodeV;
-            if (chainId == null) {
-                encodeV = signature.v;
-            } else {
-                encodeV = signature.v - LOWER_REAL_V;
-                encodeV += chainId * 2 + CHAIN_ID_INC;
-            }
-            v = RLP.encodeInt(encodeV);
-            r = RLP.encodeElement(BigIntegers.asUnsignedByteArray(signature.r));
-            s = RLP.encodeElement(BigIntegers.asUnsignedByteArray(signature.s));
-        } else {
-            // Since EIP-155 use chainId for v
-            v = chainId == null ? RLP.encodeElement(EMPTY_BYTE_ARRAY) : RLP.encodeInt(chainId);
-            r = RLP.encodeElement(EMPTY_BYTE_ARRAY);
-            s = RLP.encodeElement(EMPTY_BYTE_ARRAY);
-        }
-
-        this.rlpEncoded = RLP.encodeList(nonce, gasPrice, gasLimit,
-                receiveAddress, value, data, v, r, s);
-
-        this.hash = this.getHash();
-        */
-
-    byte[] value = Longs.toByteArray(this.value);
-    byte[] raw = new byte[this.receiveAddress.length + this.data.length + value.length];
-    System.arraycopy(this.receiveAddress, 0, raw, 0, this.receiveAddress.length);
-    System.arraycopy(this.data, 0, raw, this.receiveAddress.length, this.data.length);
-    System.arraycopy(value, 0, raw, this.data.length, value.length);
-    this.protoEncoded = raw;
-    this.hash = Hash.sha3(protoEncoded);
-
-    return protoEncoded;
-  }
-
-  @Override
-  public int hashCode() {
-
-    byte[] hash = this.getHash();
-    int hashCode = 0;
-
-    for (int i = 0; i < hash.length; ++i) {
-      hashCode += hash[i] * i;
-    }
-
-    return hashCode;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || !(o.getClass() == getClass())) {
-      return false;
-    }
-    VMTransaction tx = (VMTransaction) o;
-
-    return tx.hashCode() == this.hashCode();
-  }
-
-  /**
-   * @deprecated Use {@link VMTransaction#createDefault(String, long)} instead
-   */
-  public static VMTransaction createDefault(String to, long amount) {
-    return create(to, amount);
-  }
-
-  public static VMTransaction createDefault(String to, long amount, Integer chainId) {
-    return create(to, amount, chainId);
-  }
-
-
-  public static VMTransaction create(String to, long amount) {
-    return new VMTransaction(
-        Hex.decode(to),
-        amount,
-        null);
-  }
-
-  public static VMTransaction create(String to, long amount, Integer chainId) {
-    return new VMTransaction(
-        Hex.decode(to),
-        amount,
-        null,
-        chainId);
-  }
-}
